+++ db.json
@@ -1,4 +1,33 @@
 {
-  "blogs": {},
-  "test": 1672507149437
+  "blogs": {
+    "eb4f7706-bd5f-43e2-86bd-2291c2d27718": {
+      "title": "New Blog",
+      "description": "Some description",
+      "createdDate": 1672567766406,
+      "editedDate": 1672567766406,
+      "comments": {}
+    },
+    "2b132a90-d0cc-4047-8f03-2cc31a7d4ad7": {
+      "title": "New Blog 2",
+      "description": "Some description 2",
+      "createdDate": 1672567766415,
+      "editedDate": 1672567766415,
+      "comments": {}
+    },
+    "a971c280-7814-457e-ae7a-2bd0cfdc5961": {
+      "title": "New Blog",
+      "description": "Some description",
+      "createdDate": 1672568593437,
+      "editedDate": 1672568593437,
+      "comments": {}
+    },
+    "c727c92f-82a8-434a-bce8-5b2928b29c21": {
+      "title": "New Blog 2",
+      "description": "Some description 2",
+      "createdDate": 1672568593446,
+      "editedDate": 1672568593446,
+      "comments": {}
+    }
+  },
+  "test": 1672568592904
 }
